fix: viewPort layout으로 수정
